requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.1
cohere==4.47
djangorestframework-simplejwt==5.3.1 